Refactor gulpfile into sub-modules.
